Added barbones components for snakes page
