+++ front-end/src/app.js
@@ -1,15 +1,17 @@
 import React from 'react'
-import Menu from "./components/menu"
+import Menu from "./components/common/menu"
 import PicHolder from "./components/picHolder"
+import { Route, Switch } from 'react-router-dom'
 
 const App = () => {
     return (
-    <div className='container-fluid'>
-        <Menu />
-        
-        <PicHolder/>
-       
-    </div>
+        <div className='container-fluid'>
+            <Menu />
+            <Switch>
+                <Route exact path="/" component={PicHolder} />
+                <Route path="/snakes" />
+            </Switch>
+        </div>
     )
 }
