+++ front-end/src/components/common/Snakes.jsx
@@ -1,0 +1,10 @@
+import React from "react"
+
+const Snakes = props => {
+    return (
+        <>
+        </>
+    )
+}
+
+export default Snakes
