+++ front-end/src/components/common/SnakesPage.jsx
@@ -1,0 +1,13 @@
+import React, { useState, useEffect } from "react"
+
+const SnakePage = () => {
+    return (
+        <>
+
+        </>
+    )
+}
+
+
+
+export default SnakePage
