--- front-end/src/components/common/loadingDits.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-const LoadingDots = () => {
-    return (
-        <div class="spinner-border m-5" role="status">
-            <span class="sr-only">Analizing...</span>
-        </div>
-    )
-
-}
-
-export default LoadingDots;
